% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatisticalTests.R
\name{branch_differential_expression}
\alias{branch_differential_expression}
\title{differential expression in specific branch}
\usage{
branch_differential_expression(Branch, EmbeddedTree, mode = c("tree",
  "cells"))
}
\arguments{
\item{Branch}{It is a vector containing the indexes of cells to be part of the first Subpopulation}

\item{EmbeddedTree}{Embedded Elastic Tree as calculated by GenesSpaceEmbedding()}

\item{mode}{whether the cell expression profiles or the tree nodes expression profiles will be taken into account for the test. "cells" uses the expression values from the cells. "tree"}
}
\value{
genes_significances list object containing the p-values ($pvals) and e-values ($evals) vectors for the kruskall wallis test ordered from the highest significance to the lowest one. A vector containing the gene names ($GeneName) ordered as the significances vector is also provided. Another vector ($indexes) containing the original index order for genes in the columns from the expression matrix is also provided.
}
\description{
Given an elastic tree it calculates genes that are differentially expressed in a particular branch in the tree
}
