% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ElasticPrincipalTree.R
\name{CalculateElasticTree}
\alias{CalculateElasticTree}
\title{Calculate Elastic Tree}
\usage{
CalculateElasticTree(ScaffoldTree, N_yk, input = "topology",
  lambda_0 = 2.03e-09, mu_0 = 0.00625, FixEndpoints = F, plot = F)
}
\arguments{
\item{N_yk}{number of nodes for the elastic principal tree}

\item{lambda_0}{principal elastic tree energy function parameter.}

\item{mu_0}{principal elastic tree energy function parameter.}

\item{Scaffoldtree}{scaffoldTree calculated by the CalculateScaffoldTree function.}
}
\value{
ElasticTree
}
\description{
Calculates an elastic tree using a scaffold tree to initialize it. A set of N_yk nodes are included one byone into a tree structure that minimizes an error and an energetic functions to the cell coordinates. The initialization with the scaffold tree ensures that the correct number of endpoints and branchpoints and their connectivity are preserved in the elastic tree interpolation.
}
