% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ElasticPrincipalTree.R
\name{CalculateElasticTreeConstrained}
\alias{CalculateElasticTreeConstrained}
\title{Calculate elastic tree with iterative constraints}
\usage{
CalculateElasticTreeConstrained(ScaffoldTree, N_yk = 150, start_N_yk = 100,
  step_N_yk = 50, lambda_0 = 2.03e-09, mu_0 = 0.00625, FixEndpoints = F,
  plot = F)
}
\arguments{
\item{ScaffoldTree}{scaffoldTree calculated by the CalculateScaffoldTree function.}

\item{N_yk}{number of nodes for the elastic principal tree}

\item{start_N_yk}{initial number of nodes with which the elastic tree will be calculated on top of which additional constraints will be added}

\item{step_N_yk}{number of nodes that will be interatively added to the elastic tree until the N_yk nodes are added. After each iteration the nodes will be added as additional constraints to the next iteration.}

\item{lambda_0}{principal elastic tree energy function parameter.}

\item{mu_0}{principal elastic tree energy function parameter.}

\item{FixEndpoints}{Whether or not the end points coordinates are fixed}

\item{plot}{Whether or not plots are produced}
}
\value{
ElasticTree
}
\description{
Calculates an elastic tree using a scaffold tree to initialize it. A set of N_yk nodes are included one by one into a tree structure that minimizes an error and an energetic functions to the cell coordinates. The initialization with the scaffold tree ensures that the correct number of endpoints and branchpoints and their connectivity are preserved in the elastic tree interpolation.
}
