% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pseudotime.R
\name{CalculatePseudotimes}
\alias{CalculatePseudotimes}
\title{Calculate Pseudotime}
\usage{
CalculatePseudotimes(InputTree, plot = F, T0 = 1, C0 = NULL)
}
\arguments{
\item{InputTree}{elastic tree according to which the pseudotime will be calculated. It can be a low dimensional elastic tree as the one calculated by CalculateElasticTree() or a high dimensional elastic tree like the one calculated by GeneSpaceEmbedding()}

\item{plot}{whether or not this function will produce plots}

\item{T0}{which endpoint will be used at time zero}

\item{C0}{which cell will be used as time zero}
}
\description{
Calculates pseudotime for the cells either on the low dimensional space or the full gene expression space. Users can define which is the endpoint or cell that will be used as the initial time (t0) respect to which the pseudotimes to all cells will be calculated.
}
