% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScaffoldTree.R
\name{CalculateScaffoldTree}
\alias{CalculateScaffoldTree}
\title{Calculate Scaffold tree}
\usage{
CalculateScaffoldTree(CellCoordinates, NEndpoints = NULL,
  BranchMinLength = -1, BranchMinLengthSensitive = -1, reduced = 0,
  python_location = "python3", docker = NULL)
}
\arguments{
\item{CellCoordinates}{file containing coordinates from the low dimensional manifold (e.g: first 3 diffusion components from a diffusion map)}

\item{NEndpoints}{Users can specify how many endpoints they want the algorithm to find. In case this variable is not defined all branches producing branches longer than sqrt(N/2) will be added to the tree structure}

\item{BranchMinLength}{Minimum number of nodes a branch has to contain in order to be included in the 3 structure. By default this value is set to sqrt(N/2) with N being the total number of cells in the dataset.}

\item{BranchMinLengthSensitive}{Minimum length for a branch to be included in the tree. It reconstructs the topology of the tree and maps cells to the potential new branch to decide if the branch will be added or not. Suggested value: sqrt(N) with N being the number of cells in the dataset}

\item{reduced}{The number of clusters to group cells in. If set to 0, no clustering will be performed and the scaffold tree will be calculated on all cells (default).}

\item{python_location}{url to the python3 executable binary. In case it is not specified a default call to python3 will be used.}

\item{docker}{name of the docker container that accompanies MERLoT (alternative to using a local python installation).}
}
\value{
ScaffoldTree object with the structure and connectivity of the Scaffold Tree.
}
\description{
creates the scaffold tree from the expression matrix in the Dataset object.
}
