% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiffusionMap.R
\name{dataset_n_observations}
\alias{dataset_n_observations}
\title{From destiny: Helper function for diffusion map calculations.}
\usage{
dataset_n_observations(data, distances)
}
\arguments{
\item{data}{The data matrix}

\item{distances}{The matrix of pairwise distances of cells.}
}
\value{
The number of observations in the data
}
\description{
From destiny: Helper function for diffusion map calculations.
}
