% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ElasticPrincipalTree.R
\name{DuplicateTreeNodes}
\alias{DuplicateTreeNodes}
\title{Duplicate Elastic Tree Nodes
Introduces intermediate nodes in between nodes being part of an edge in the ElasticTree structure}
\usage{
DuplicateTreeNodes(ElasticTree)
}
\arguments{
\item{ElasticTree}{Elastic Tree to which nodes will be added}
}
\value{
ElasticTree
}
\description{
Duplicate Elastic Tree Nodes
Introduces intermediate nodes in between nodes being part of an edge in the ElasticTree structure
}
