% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ElasticPrincipalTree.R
\name{GenesSpaceEmbedding}
\alias{GenesSpaceEmbedding}
\title{Gene Space Embedding}
\usage{
GenesSpaceEmbedding(ExpressionMatrix, ElasticTree, lambda_0 = 2.03e-09,
  mu_0 = 0.00625, increaseFactor_mu = 20, increaseFactor_lambda = 20,
  NCores = 1)
}
\arguments{
\item{ExpressionMatrix}{Gene expression matrix on which the tree will be embedded}

\item{ElasticTree}{elastic tree to be embedded on the gene expression space}

\item{increaseFactor}{factor by which the principal elastic tree energy function parameters will be increased for the embedding. By default this number is 10.}
}
\description{
Embeds an elastic principal tree on the gene expression space from which the initial cell coordinates were calculated. The function receives the original Expression Matrix from which a lowe dimensional manifold was calculated and the coordinates from the low-dimensional manifold.
}
