% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotTree.R
\name{get_node_graph}
\alias{get_node_graph}
\title{Calculates an adjacency matrix for the support nodes of an elastic tree and returns the corresponding graph.}
\usage{
get_node_graph(ElasticTree)
}
\arguments{
\item{ElasticTree}{the tree to convert}
}
\value{
igraph graph
}
\description{
Calculates an adjacency matrix for the support nodes of an elastic tree and returns the corresponding graph.
}
