% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatisticalTests.R
\name{GetGeneCorrelationNetwork}
\alias{GetGeneCorrelationNetwork}
\title{Get Gene Correlation Network}
\usage{
GetGeneCorrelationNetwork(ExpressionMatrix, cor_threshold = 0.4,
  plot = T)
}
\arguments{
\item{ExpressionMatrix}{contaings the expression profiles(columns) along the cells (rows)}

\item{cor_threshold}{correlation threshold for showing edges in the network plot (default=0.4)}

\item{plot}{whether the network should be plotted or not (default=T)}
}
\value{
cor.genes pair correlation matrix
}
\description{
Given an gene expression matrix (raw or embedded in the high dimensional elastic tree) it calculates the correlations in between all pairs of gene expression profiles and plots the network
}
