% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ElasticPrincipalTree.R
\name{GetTrajectory}
\alias{GetTrajectory}
\title{Get Cells from Trajectory}
\usage{
GetTrajectory(ElasticTree, Start, End, Pseudotimes = NULL)
}
\arguments{
\item{ElasticTree}{Elastic tree from which cells will be taken}

\item{Start}{element in the topology from where the trajectory starts}

\item{End}{element in the topology where the trajectory ends}

\item{Pseudotimes}{pseudotime object to provide also the pseudotimes for the cells and nodes in the trajectory}
}
\value{
Trajectory
}
\description{
Selects the cells that belong to a trajectory in an Elastric Tree between two elements in the tree topology.
}
