% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ElasticPrincipalTree.R
\name{inflate_elastic_tree}
\alias{inflate_elastic_tree}
\title{Takes an elastic tree calculated on reduced coordinates and replaces the clusters with the real cells.}
\usage{
inflate_elastic_tree(ElasticTree, FullCoordinates)
}
\arguments{
\item{ElasticTree}{the elastic tree calculated by the CalculateElasticTree function with a reduced input from ScaffoldTree}

\item{FullCoordinates}{the manifold coordinates of the cells}
}
\value{
ElasticTree
}
\description{
This function will replace the CellCoords field of the elastic tree and reassign cells to branch nodes and branches.
}
