% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadDataset.R
\name{LoadExprMatrix}
\alias{LoadExprMatrix}
\title{LoadExprMatrix}
\usage{
LoadExprMatrix(ExprMatrix)
}
\arguments{
\item{ExprMatrix}{matrix object with n rows as cells and m columns as genes.}
}
\value{
Dataset an object with the Expression Matrix, cell names and gene names
}
\description{
LoadExprMatrix
}
