% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotTree.R
\name{plot_elastic_tree}
\alias{plot_elastic_tree}
\title{Plot Elastic Tree}
\usage{
plot_elastic_tree(ElasticTree, colorcells = NULL, legend = T)
}
\arguments{
\item{ElasticTree}{Principal Elastic Tree calculated with the CalculateElasticTreeFunction}

\item{colorcells}{vector of colors for the cells}
}
\description{
Plots an elastic tree on top of the cell coordinates on which it was calculated
}
