% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotTree.R
\name{plot_flattened_tree}
\alias{plot_flattened_tree}
\title{Plots a flattened 2 dimensional version of the Elastic Tree Nodes}
\usage{
plot_flattened_tree(ElasticTree, cell_annot = NULL, layout = NULL,
  legend_position = "topright", node_size = c("topology", "cells"),
  node_radius = 12)
}
\arguments{
\item{ElasticTree}{One of the elements in the Dataset$GeneNames object.}

\item{legend_position}{The position of the legend.}
}
\description{
Given an Elastic Tree it plots a 2 dimensional version of the Elastic Tree.
}
