% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotTree.R
\name{plot_flattened_tree_gene_expression}
\alias{plot_flattened_tree_gene_expression}
\title{Plots the gene expression level of a gene on a flattened 2 dimensional version of the Elastic Tree Nodes}
\usage{
plot_flattened_tree_gene_expression(ElasticTree, GeneName,
  legend_position = "topright")
}
\arguments{
\item{ElasticTree}{One of the elements in the Dataset$GeneNames object.}

\item{GeneName}{Name of the gene to be visualized. The gene name is one of the colnames in the EmbeddedTree$Nodes object}

\item{legend_position}{The position of the legend.}
}
\description{
Given an Elastic Tree it plots a 2 dimensional version of the Elastic Tree.
}
