% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotTree.R
\name{plot_gene_on_map}
\alias{plot_gene_on_map}
\title{Plots Gene Expression on Cells Coordinates}
\usage{
plot_gene_on_map(GeneName, CellCoordinates, ExpressionMatrix,
  CoordLabels = c("Component 1", "Component 2", "Component 3"),
  Average = F, knn = 5, Palette = grDevices::colorRampPalette(colors
  = c("yellow", "orange", "red", "black")))
}
\arguments{
\item{GeneName}{One of the elements in the Dataset$GeneNames object.}

\item{CellCoordinates}{Matrix with 2 or 3 coordinates per cell.}

\item{ExpressionMatrix}{Expression matrix}

\item{CoordLabels}{Which coordinate labels to use}

\item{Average}{wheter or not expression for cells will be averaged over the nearest knn neighbors. Default Average=FALSE}

\item{knn}{number of neighbors over which expression for the cells will be averaged in case Average=TRUE. Default: knn=5}

\item{Palette}{colors that will be used for the cells by default= Palette=colorRampPalette(colors=c("yellow", "orange", "red","black")))}
}
\description{
Plots the expression of a given gene on a given set of coordinates
}
