% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotTree.R
\name{plot_pseudotimes}
\alias{plot_pseudotimes}
\title{Plot cells pseudotime}
\usage{
plot_pseudotimes(CellCoordinates, Pseudotimes)
}
\arguments{
\item{CellCoordinates}{matrix containing up to 3 coordinates for cells in the manifold space}

\item{Pseudotimes}{object calculated by CalculatePseudotimes()}
}
\description{
Plot pseudotime for the cells in the given cell coordinatges
}
