% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotTree.R
\name{plot_scaffold_tree}
\alias{plot_scaffold_tree}
\title{Plot Scaffold Tree}
\usage{
plot_scaffold_tree(ScaffoldTree, colorcells = "gray",
  dims = dim(ScaffoldTree$CellCoordinates)[2])
}
\arguments{
\item{ScaffoldTree}{scaffoldTree calculated by the CalculateScaffoldTree function.}

\item{colorcells}{The colours to use for the cells}

\item{dims}{The number of dimensions to plot}
}
\description{
Plots the scaffold tree on top of the distribution of cells in 2 or 3 coordinates.
}
