% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadDataset.R
\name{ReadDataset}
\alias{ReadDataset}
\title{ReadDataset}
\usage{
ReadDataset(file, featuresfile = c())
}
\arguments{
\item{file}{filename including the path to where the expression matrix is contained. The file needs to have the cells as rows and the genes as columns.}

\item{featuresfile}{optional file with features for the different cells in the dataset}
}
\value{
Dataset an object with the Expression Matrix, cell names and gene names
}
\description{
Read the dataset from a file containing gene the expression matrix
}
