% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneSelection.R
\name{seurat_normalize}
\alias{seurat_normalize}
\title{seurat normalize}
\usage{
seurat_normalize(ExprMatrix, scale.factor = 10000)
}
\arguments{
\item{ExprMatrix}{cells x genes expression matrix to be normalized}

\item{scale.factor}{scaling factor to be used for normalizing the data}
}
\description{
Normalizes an expression matrix following the procedure as implemented in the Seurat's package
}
