% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneSelection.R
\name{seurat_variable_genes}
\alias{seurat_variable_genes}
\title{Seurat variable genes, according to the Seurat's package}
\usage{
seurat_variable_genes(ExprMatrix, bin.size = 1000, x.low.cutoff = 0.1,
  x.high.cutoff = 8, y.cutoff = 1, y.high.cutoff = Inf, num.bin = 20)
}
\arguments{
\item{ExprMatrix}{cells x genes expression matrix to be normalized}

\item{bin.size}{The bin size to use for the gene histogram}

\item{x.low.cutoff}{Bottom cutoff on x-axis for identifying variable genes}

\item{x.high.cutoff}{Top cutoff on x-axis for identifying variable genes}

\item{y.cutoff}{Bottom cutoff on y-axis for identifying variable genes}

\item{y.high.cutoff}{Top cutoff on y-axis for identifying variable genes}

\item{num.bin}{Total number of bins to use in the scaled analysis (default is 20)}
}
\value{
ScaffoldTre object with the structure and connectivity of the Scaffold Tree
}
\description{
Identifies genes that are outliers on a 'mean variability plot'. First, calculates
average expression and dispersion for each gene. Next, divides genes into
num.bin (deafult 20) bins based on
their average expression, and calculates z-scores for dispersion within each
bin. The purpose of this is to identify variable genes while controlling for
the strong relationship between variability and average expression.
}
\details{
Exact parameter settings may vary empirically from dataset to dataset, and
based on visual inspection of the plot.
Setting the y.cutoff parameter to 2 identifies genes that are more than two standard
deviations away from the average dispersion within a bin. The default X-axis function
is the mean expression level, and for Y-axis it is the log(Variance/mean). All mean/variance
calculations are not performed in log-space, but the results are reported in log-space -
see relevant functions for exact details.
}
