% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScaffoldTree.R
\name{CalculateScaffoldTree}
\alias{CalculateScaffoldTree}
\title{Calculate Scaffold tree}
\usage{
CalculateScaffoldTree(CellCoordinates, NEndpoints = NULL,
  python_location = "python3")
}
\arguments{
\item{CellCoordinates}{file containing coordinates from the low dimensional manifold (e.g: first 3 diffusion components from a diffusion map)}

\item{NEndpoints}{Users can specify how many endpoints they want the algorithm to find. In case this variable is not defined all branches producing branches longer than sqrt(N/2) will be added to the tree structure}

\item{python_location}{url to the python3 executable binary. In case it is not specified a default call to python3 will be used.}
}
\value{
ScaffoldTre object with the structure and connectivity of the Scaffold Tree
}
\description{
creates the scaffold tree from the expression matrix in the Dataset object.
}
