% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotTree.R
\name{plot_flattened_tree}
\alias{plot_flattened_tree}
\title{Plots a flattened 2 dimensional version of the Elastic Tree Nodes}
\usage{
plot_flattened_tree(ElasticTree)
}
\arguments{
\item{ElasticTree}{One of the elements in the Dataset$GeneNames object.}
}
\description{
Given an Elastic Tree it plots a 2 dimensional version of the Elastic Tree.
}
