% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotTree.R
\name{plot_scaffold_tree}
\alias{plot_scaffold_tree}
\title{Plot Scaffold Tree}
\usage{
plot_scaffold_tree(ScaffoldTree, colorcells = "gray",
  dims = dim(ScaffoldTree$CellCoordinates)[2])
}
\arguments{
\item{CellCoordinates}{coordinates for the cells. This are the same coordinates that were used in the Scaffold Tree Calculation}
}
\description{
Plots the scaffold tree on top of the distribution of cells in 2 or 3 coordinates.
}
